requirements.txt
django
djangorestframework
gunicorn
mysqlclient==1.4.1
coverage
django-heroku
django-cors-headers